fix(cr): fix cr comments
